requirements.txt
python-dotenv
requests==2.30.0
fastapi==0.95.2
uvicorn==0.22.0
pinecone-client==2.2.1
openai==0.27.7
sentence-transformers==2.2.2
tiktoken